requirements.txt
fastmcp
markitdown
fastapi
uvicorn
pydantic
requests
boto3
python-dotenv
click
ddgs
pydantic-ai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
authlib
litellm
itsdangerous
starlette
pydantic-ai
fastapi-mcp

